--- src/main/java/com/cherryhouse/server/style/StyleService.java
@@ -1,4 +1,0 @@
-package com.cherryhouse.server.style;
-
-public class StyleService {
-}
