+++ src/main/java/com/cherryhouse/server/user/UserService.java
@@ -6,14 +6,13 @@ import com.cherryhouse.server._core.util.PageData;
 import com.cherryhouse.server.auth.dto.AuthRequest;
 import com.cherryhouse.server.post.Post;
 import com.cherryhouse.server.post.PostRepository;
-import com.cherryhouse.server.post.PostService;
 import com.cherryhouse.server.post.image.ImageMapping;
 import com.cherryhouse.server.post.image.ImageService;
 import com.cherryhouse.server.post.posttag.PostTagMapping;
 import com.cherryhouse.server.post.tag.TagService;
 import com.cherryhouse.server.s3.S3Service;
-import com.cherryhouse.server.style.Style;
-import com.cherryhouse.server.style.StyleRepository;
+import com.cherryhouse.server.user.style.Style;
+import com.cherryhouse.server.user.style.StyleRepository;
 import com.cherryhouse.server.user.dto.UserRequest;
 import com.cherryhouse.server.user.dto.UserResponse;
 import lombok.RequiredArgsConstructor;
