+++ src/main/java/com/cherryhouse/server/user/dto/UserResponse.java
@@ -4,7 +4,7 @@ import com.cherryhouse.server._core.util.PageData;
 import com.cherryhouse.server.post.Post;
 import com.cherryhouse.server.post.image.ImageMapping;
 import com.cherryhouse.server.post.posttag.PostTagMapping;
-import com.cherryhouse.server.style.Style;
+import com.cherryhouse.server.user.style.Style;
 import com.cherryhouse.server.user.User;
 
 import java.util.List;
