+++ src/main/java/com/cherryhouse/server/user/style/StyleRepository.java
@@ -1,4 +1,4 @@
-package com.cherryhouse.server.style;
+package com.cherryhouse.server.user.style;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
