+++ src/main/java/com/cherryhouse/server/user/style/StyleService.java
@@ -1,0 +1,4 @@
+package com.cherryhouse.server.user.style;
+
+public class StyleService {
+}
